remove handler package
