+++ api/src/main/java/com/pfplaybackend/api/config/websocket/event/listener/ConnectionEventListener.java
@@ -1,4 +1,4 @@
-package com.pfplaybackend.api.config.websocket.event;
+package com.pfplaybackend.api.config.websocket.event.listener;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
