+++ api/src/main/java/com/pfplaybackend/api/config/websocket/event/manager/SessionCacheManager.java
@@ -1,4 +1,4 @@
-package com.pfplaybackend.api.config.websocket.event.handler;
+package com.pfplaybackend.api.config.websocket.event.manager;
 import com.pfplaybackend.api.user.domain.value.UserId;
 
 public interface SessionCacheManager {
